scraper/requirements.txt
# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# HTTP Requests
httpx==0.26.0
aiohttp==3.9.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80  # Computer Vision for image quality assessment

# Data Handling
pydantic==2.6.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.1
python-slugify==8.0.3
validators==0.22.0

# Logging & Monitoring
loguru==0.7.2
structlog==24.1.0

# Async
asyncio==3.4.3
aiofiles==23.2.1

# CLI
click==8.1.7
rich==13.7.0

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
